# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/v1/products": {
      "post": {
        "operationId": "ProductsAdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Create a product (with optional variants)",
        "tags": [
          "Admin / Products"
        ]
      },
      "get": {
        "operationId": "ProductsAdminController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 20,
              "type": "integer"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minLength": 1,
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "published",
                "archived"
              ]
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "tag",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "priceMin",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "priceMax",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "type": "integer"
            }
          },
          {
            "name": "inStock",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "default": "created",
              "type": "string",
              "enum": [
                "created",
                "title",
                "price"
              ]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "x-nestjs_zod-parent-additional-properties": false,
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List products (offset pagination, filters, sort)",
        "tags": [
          "Admin / Products"
        ]
      }
    },
    "/admin/v1/products/{id}": {
      "get": {
        "operationId": "ProductsAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get product by ID with variants and images",
        "tags": [
          "Admin / Products"
        ]
      },
      "patch": {
        "operationId": "ProductsAdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update product (PATCH semantics)",
        "tags": [
          "Admin / Products"
        ]
      },
      "delete": {
        "operationId": "ProductsAdminController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Hard delete product (CASCADE removes variants/images)",
        "tags": [
          "Admin / Products"
        ]
      }
    },
    "/admin/v1/products/{id}/images": {
      "post": {
        "operationId": "ProductsAdminController_attachImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachImageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Attach an uploaded image to a product",
        "tags": [
          "Admin / Products"
        ]
      }
    },
    "/admin/v1/products/{id}/images/{imageId}": {
      "delete": {
        "operationId": "ProductsAdminController_detachImage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "imageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Detach an image from a product",
        "tags": [
          "Admin / Products"
        ]
      }
    },
    "/admin/v1/products/{id}/images/reorder": {
      "post": {
        "operationId": "ProductsAdminController_reorderImages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderImagesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Reorder product images",
        "tags": [
          "Admin / Products"
        ]
      }
    },
    "/admin/v1/products/{id}/categories": {
      "put": {
        "operationId": "ProductsAdminController_assignCategories",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCategoriesDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Replace product category set (replace-set semantics)",
        "tags": [
          "Admin / Products"
        ]
      }
    },
    "/admin/v1/products/{id}/tags": {
      "put": {
        "operationId": "ProductsAdminController_assignTags",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTagsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Replace product tag set (replace-set semantics)",
        "tags": [
          "Admin / Products"
        ]
      }
    },
...
    "/store/v1/products/{slug}": {
      "get": {
        "operationId": "ProductsStoreController_findOne",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get a published product by slug (no auth)",
        "tags": [
          "Store / Products"
        ]
      }
    },
...
  },
  "info": {
    "title": "SovEcom API",
    "description": "The SovEcom headless ecommerce API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "status": {
            "default": "draft",
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "seoTitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "seoDescription": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "isBundle": {
            "default": false,
            "type": "boolean",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "variants": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "sku": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 255
                },
                "title": {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                "options": {
                  "default": {},
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                },
                "priceAmount": {
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "currency": {
                  "type": "string",
                  "pattern": "^[A-Za-z]{3}$"
                },
                "compareAtAmount": {
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "stockQuantity": {
                  "default": 0,
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                },
                "allowBackorder": {
                  "default": false,
                  "type": "boolean"
                },
                "weightGrams": {
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "lengthMm": {
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "widthMm": {
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "heightMm": {
                  "anyOf": [
                    {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "position": {
                  "default": 0,
                  "type": "integer",
                  "minimum": 0,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "priceAmount",
                "currency"
              ],
              "additionalProperties": false
            },
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "maxLength": 512,
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published",
              "archived"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "seoTitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "seoDescription": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1024
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true,
            "x-nestjs_zod-parent-additional-properties": false
          },
          "isBundle": {
            "type": "boolean",
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-parent-additional-properties": false
          }
        }
      },
      "AttachImageDto": {
        "type": "object",
        "properties": {
          "imageId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
            "x-nestjs_zod-parent-additional-properties": false
          },
          "position": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 10000,
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "imageId"
        ]
      },
      "ReorderImagesDto": {
        "type": "object",
        "properties": {
          "order": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "order"
        ]
      },
      "AssignCategoriesDto": {
        "type": "object",
        "properties": {
          "categoryIds": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "categoryIds"
        ]
      },
      "AssignTagsDto": {
        "type": "object",
        "properties": {
          "tagIds": {
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "x-nestjs_zod-parent-additional-properties": false
          }
        },
        "required": [
          "tagIds"
        ]
      },
...
    }
  }
}